backend/requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
boto3==1.16.38
botocore==1.19.38
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
Django==3.1.4
django-cacheops==5.1
django-cors-headers==3.6.0
django-extensions==3.1.1
django-redis==4.12.1
django-seed==0.2.2
django-sslserver==0.22
django-storages==1.11
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-nested-routers==0.93.3
drf-yasg==1.20.0
Faker==5.0.2
funcy==1.15
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
kombu==5.0.2
MarkupSafe==1.1.1
packaging==20.9
Pillow==8.1.0
prompt-toolkit==3.0.17
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
